# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bipupu",
    "description": "BIPUPU API 服务",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/service_accounts/": {
      "get": {
        "tags": [
          "服务号",
          "服务号"
        ],
        "summary": "List Service Accounts",
        "description": "获取所有活跃的服务号列表\n\n参数：\n- skip: 跳过的记录数，用于分页\n- limit: 每页返回的记录数，最大100\n\n返回：\n- items: 服务号列表\n- total: 活跃服务号总数\n\n注意：只返回 is_active=True 的服务号",
        "operationId": "list_service_accounts_api_service_accounts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAccountList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/service_accounts/{name}": {
      "get": {
        "tags": [
          "服务号",
          "服务号"
        ],
        "summary": "Get Service Account",
        "description": "获取指定服务号详情\n\n参数：\n- name: 服务号名称\n\n返回：\n- 成功：返回服务号详细信息\n- 失败：404（服务号不存在）\n\n注意：需要用户认证",
        "operationId": "get_service_account_api_service_accounts__name__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/service_accounts/{name}/avatar": {
      "get": {
        "tags": [
          "服务号",
          "服务号"
        ],
        "summary": "Get Service Avatar",
        "description": "获取服务号头像\n\n参数：\n- name: 服务号名称\n\n特性：\n- 支持ETag缓存，减少带宽消耗\n- 支持HTTP 304 Not Modified响应\n- 头像数据缓存24小时\n\n返回：\n- 成功：返回JPEG格式的头像图片\n- 失败：404（服务号或头像不存在）",
        "operationId": "get_service_avatar_api_service_accounts__name__avatar_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/service_accounts/subscriptions/": {
      "get": {
        "tags": [
          "服务号",
          "订阅"
        ],
        "summary": "Get User Subscriptions",
        "description": "获取当前用户订阅的所有服务号列表\n\n返回：\n- subscriptions: 订阅的服务号列表，包含服务信息和订阅设置\n- total: 订阅的服务号总数\n\n包含信息：\n- 服务号基本信息（名称、描述、头像等）\n- 订阅设置（推送时间、启用状态、订阅时间等）\n\n注意：需要用户认证",
        "operationId": "get_user_subscriptions_api_service_accounts_subscriptions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSubscriptionList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/service_accounts/{name}/settings": {
      "get": {
        "tags": [
          "服务号",
          "订阅"
        ],
        "summary": "Get Subscription Settings",
        "description": "获取指定服务号的订阅设置\n\n参数：\n- name: 服务号名称\n\n返回：\n- 成功：返回该服务号的订阅设置详情\n- 失败：404（服务号不存在或未订阅）\n\n包含信息：\n- 服务号基本信息\n- 推送时间设置及来源\n- 订阅启用状态\n- 订阅时间戳\n\n注意：需要用户认证且已订阅该服务号",
        "operationId": "get_subscription_settings_api_service_accounts__name__settings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "服务号",
          "订阅"
        ],
        "summary": "Update Subscription Settings",
        "description": "更新服务号订阅设置\n\n参数：\n- name: 服务号名称\n- push_time: 推送时间（HH:MM格式），可选\n- is_enabled: 是否启用推送，可选\n\n返回：\n- 成功：返回更新后的订阅设置\n- 失败：404（服务号不存在或未订阅）或400（时间格式无效）\n\n注意：\n- 需要用户认证且已订阅该服务号\n- 推送时间格式必须为 HH:MM（24小时制）\n- 如果只更新部分字段，其他字段保持不变\n- 设置 push_time=null 可清除个人化设置，恢复使用默认时间",
        "operationId": "update_subscription_settings_api_service_accounts__name__settings_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionSettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/service_accounts/{name}/subscribe": {
      "post": {
        "tags": [
          "服务号",
          "订阅"
        ],
        "summary": "Subscribe Service Account",
        "description": "订阅服务号\n\n参数：\n- name: 要订阅的服务号名称\n- push_time: 初始推送时间（HH:MM格式），可选\n- is_enabled: 初始启用状态，可选，默认为True\n\n返回：\n- 成功：返回订阅成功信息\n- 失败：404（服务号不存在）或400（已订阅）或500（数据库操作失败）\n\n注意：\n- 需要用户认证\n- 只能订阅 is_active=True 的服务号\n- 推送时间格式必须为 HH:MM（24小时制）\n- 如果未提供推送时间，使用服务号默认推送时间",
        "operationId": "subscribe_service_account_api_service_accounts__name__subscribe_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionSettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "服务号",
          "订阅"
        ],
        "summary": "Unsubscribe Service Account",
        "description": "取消订阅服务号\n\n参数：\n- name: 要取消订阅的服务号名称\n\n返回：\n- 成功：返回取消订阅成功信息\n- 失败：404（服务号不存在）或400（未订阅）或500（数据库操作失败）\n\n注意：\n- 需要用户认证\n- 取消订阅后，该服务号的推送将停止\n- 订阅记录将被删除，相关设置不会保留",
        "operationId": "unsubscribe_service_account_api_service_accounts__name__subscribe_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/service_accounts": {
      "get": {
        "tags": [
          "管理后台"
        ],
        "summary": "Admin Services",
        "description": "服务号管理页面",
        "operationId": "admin_services_admin_service_accounts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/service_accounts/{service_id}/avatar": {
      "post": {
        "tags": [
          "管理后台"
        ],
        "summary": "Upload Service Avatar",
        "description": "上传服务号头像",
        "operationId": "upload_service_avatar_admin_service_accounts__service_id__avatar_post",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Service Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_service_avatar_admin_service_accounts__service_id__avatar_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_service_avatar_admin_service_accounts__service_id__avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_service_avatar_admin_service_accounts__service_id__avatar_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceAccountList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ServiceAccountResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ServiceAccountList"
      },
      "ServiceAccountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at"
        ],
        "title": "ServiceAccountResponse"
      },
...
      "SubscriptionSettingsResponse": {
        "properties": {
          "push_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Push Time",
            "description": "推送时间，格式: HH:MM"
          },
          "is_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enabled",
            "description": "是否启用推送",
            "default": true
          },
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "service_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Description"
          },
          "subscribed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Subscribed At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "push_time_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Push Time Source",
            "description": "推送时间来源: subscription(订阅设置), service_default(服务号默认), none(无)"
          }
        },
        "type": "object",
        "required": [
          "service_name",
          "service_description",
          "subscribed_at",
          "updated_at"
        ],
        "title": "SubscriptionSettingsResponse",
        "description": "订阅设置响应"
      },
...
      "UserSubscriptionList": {
        "properties": {
          "subscriptions": {
            "items": {
              "$ref": "#/components/schemas/UserSubscriptionResponse"
            },
            "type": "array",
            "title": "Subscriptions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "subscriptions",
          "total"
        ],
        "title": "UserSubscriptionList",
        "description": "用户订阅列表响应"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}